refactor(follow-list): header adjust
